app/requirements.txt
requests
beautifulsoup4
pyyaml
feedparser